sepolia_pipeline/client/requirements.txt
opencv-python
pycryptodome
requests
numpy
